samples/chat-agent/pyproject.toml
[project]
name = "chat-agent"
version = "0.0.1"
description = "Chat agent with Tavily web search"
authors = [{ name = "John Doe", email = "john.doe@myemail.com" }]
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.4",
    "langchain-anthropic>=1.2.0",
    "tavily-python>=0.7.13",
    "langchain-tavily>=0.2.13",
    "uipath-langchain>=0.1.0, <0.2.0",
    "uipath>=2.2.0, <2.3.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[dependency-groups]
dev = [
    "uipath-dev>=0.0.5",
]
